some fixes

+++ src/main/java/com/study/dao/ProductDao.java
@@ -12,6 +12,5 @@ public interface ProductDao {
 
     void updateProduct(Product product);
 
-    void deleteProduct(int id);
-
+    void deleteProduct(long id);
 }
